requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pytest==7.3.1
httpx==0.24.0
flake8==6.0.0
black==23.3.0
strawberry-graphql==0.206.0
pytest-cov
